Create authentication views

+++ blog_app/blog/views.py
@@ -1,3 +1,7 @@
+from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
+from django.http import HttpResponseForbidden
+from django.contrib.auth import login, logout
+from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect, get_object_or_404
 from .models import Post
 from .forms import PostForm
@@ -6,10 +10,12 @@ from .forms import PostForm
 def home(request):
     return render(request, 'base.html')
 
+@login_required
 def post_list(request):
     posts = Post.objects.all().order_by('-created_at')
     return render(request, 'blog/blog_list.html', {'posts': posts})
 
+@login_required
 def create_post(request):
     if request.method == 'POST':
         form = PostForm(request.POST)
@@ -22,8 +28,12 @@ def create_post(request):
         form = PostForm()
     return render(request, 'blog/create_post.html', {'form': form})
 
+@login_required
 def update_post(request, post_id):
     post = get_object_or_404(Post, id=post_id)
+    if post.author != request.user:
+        return HttpResponseForbidden("You are not allowed to edit this post.")
+    
     if request.method == 'POST':
         form = PostForm(request.POST, instance=post)
         if form.is_valid():
@@ -33,9 +43,39 @@ def update_post(request, post_id):
         form = PostForm(instance=post)
     return render(request, 'blog/create_post.html', {'form': form})
 
+@login_required
 def delete_post(request, post_id):
     post = get_object_or_404(Post, id=post_id)
+    if post.author != request.user:
+        return HttpResponseForbidden("You are not allowed to delete this post.")
+    
     if request.method == 'POST':
         post.delete()
         return redirect('post_list')
     return render(request, 'blog/delete_post.html', {'post': post})
+
+def signup_view(request):
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save()
+            login(request, user)
+            return redirect('post_list')
+    else:
+        form = UserCreationForm()
+    return render(request, 'auth/signup.html', {'form': form})
+
+def login_view(request):
+    if request.method == 'POST':
+        form = AuthenticationForm(request, data=request.POST)
+        if form.is_valid():
+            user = form.get_user()
+            login(request, user)
+            return redirect('post_list')
+    else:
+        form = AuthenticationForm()
+    return render(request, 'auth/login.html', {'form': form})
+
+def logout_view(request):
+    logout(request)
+    return redirect('login')
